error collection

+++ app/Http/Controllers/TasksController.php
@@ -74,12 +74,15 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) {
+            $user = \Auth::user();
         
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
-        
+            $task = Task::findOrFail($id);
+            
+            return view('tasks.show', [
+                'task' => $task,
+            ]);
+        }
         return redirect('/');
     }
 
@@ -91,11 +94,16 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            
+            $task = Task::findOrFail($id);
+            
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
         
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        }
        
     }
 
@@ -108,15 +116,20 @@ class TasksController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $request->validate([
-           'status' => 'required|max:10', 
-        ]);
+        if (\Auth::check()) {
+            $user = \Auth::user();
         
-        $task = Task::findOrFail($id);
+            $request->validate([
+               'status' => 'required|max:10', 
+            ]);
+            
+            $task = Task::findOrFail($id);
+            
+            $task->content = $request->content;
+            $task->status = $request->status;
+            $task->save();
         
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
+        }
         
         return redirect('/');
     }
@@ -129,10 +142,15 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        $task = Task::findOrFail($id);
+        if (\Auth::check()) {
+            $user = \Auth::user();
+            
+            $task = Task::findOrFail($id);
+            
+            if (\Auth::id() === $task->user_id) {
+                $task->delete();
+            }
         
-        if (\Auth::id() === $task->user_id) {
-            $task->delete();
         }
         
         return redirect('/');
